♻️ [Refactor] popularProducts component
